+++ cinema-main-service/src/main/java/ru/cinema/service/FileService.java
@@ -11,5 +11,5 @@ import java.util.Optional;
  * @since 15.02.2023
  */
 public interface FileService {
-    Optional<FileDto> getFileById(int id);
+    Optional<FileDto> getFileById(long id);
 }
