+++ cinema-main-service/src/main/java/ru/cinema/service/GenreService.java
@@ -11,5 +11,5 @@ import java.util.Optional;
  * @since 15.02.2023
  */
 public interface GenreService {
-    Optional<Genre> getGenreById(int id);
+    Optional<Genre> getGenreById(long id);
 }
